Language.Core

Types

data Term

Constructors

Free String
Free variable

Lambda String Term
Lambda abstraction

Con String [Term]
Constructor

Apply Term Term
Application

Fun String
Function call

Case Term [Branch]
Case statement

Let String Term Term
Let abstraction

Letrec String Term Term
Recursive let abstractoni

Unfold String Term
Unfolding

Label String Term
Label

Subst Term Term
Substitution

data Branch

Constructors

Branch String [String] Term
Standard case constructor branch (case ... of String [String] -> Term)

data Program

Constructors

Program ModuleName [ModulePragma] (Maybe WarningText) (Maybe [ExportSpec]) [ImportDecl] Term Term [(String, Term)]

Functions

parseFile :: FilePath -> IO Program

parseFile reads in a file, and converts it to a Program.

parseToCore :: Module -> Program

parseToCore accepts a Module and converts it to a Program