Language.Core.Parser

Documentation

parseFile :: FilePath -> IO Program

parseFile reads in a file, and converts it to a Program.

parseToCore :: Module -> Program

parseToCore accepts a Module and converts it to a Program